Updated the overall design for the Knative website.   (#49)
* Initial Design Updates

Simplified the colors and constrast, fixed the grid and alignment and updated the typography a bit. I've also changed the buttons to a flat and rect style which feels more modern and works better with the angular nature of the Knative logo.

* Updated the Slack/Github/Twitter area on the home page.

* Added baseof.html (overriding Docsy)

Shrunk and recolored the RSS button.

* Updated font-sizing with the documentation areas

Reverted to original sizing within the docs area, the larger fonts still apply at the top levels.

* Deleted public and resources folders and rebuilt with 'hugo' command to fix netlify extended version issues.

* Remove temporary popup modal (kubecon survey link)

* Create _sidebar-tree.scss

Added top padding for the sidebar search box.

* delete public/resources and rebuild.

Testing sidebar change

* Create _sidebar-toc.scss

Fixed the spacing and sizing on the right sidebar TOC. Moved away from the Github controls.

* Added resource folders.

* Tweaked the breadcrumb top-padding

For alignment with the sidebar content.

* Moved the rss button down for better alignment

* Added a viusal indicator to main nav of currently seleceted item

Top border on the current item.

* Changed footer Z-index to hide left nav scroll bar

* Added 'active' navbar indicator to documentation pages

* Added the active color to the left nav in docs

* Added break-word to table links

Makes tables wrap better as screen size gets smaller.